requirements.txt
ultralytics~=8.3.85
opencv-python~=4.11.0.86
fastapi~=0.115.11
uvicorn
python-dotenv~=1.0.1
sqlalchemy~=2.0.38
alembic~=1.15.1
pydantic
starlette~=0.46.0
python-multipart
numpy~=2.1.1